refactor(tests): Tier 3 final sweep — split 7 remaining oversized test files
Split 7 test files that exceeded the 250 non-comment line / 8+ test function threshold:

- internal/prereq/checker_test.go (312 lines, 19 tests)
  → checker_docker_test.go, checker_crossarch_test.go,
    checker_validate_test.go, checker_stages_test.go, checker_internals_test.go

- internal/runner/runner_test.go (306 lines, 16 tests)
  → runner_env_test.go, runner_exec_test.go, runner_output_test.go

- internal/ddc/ddc_test.go (304 lines, 17 tests)
  → ddc_path_test.go, ddc_size_test.go, ddc_ops_test.go, ddc_helpers_test.go

- internal/cleanup/cleanup_test.go (279 lines, mock-heavy)
  → cleanup_ecr_test.go, cleanup_s3_test.go

- internal/status/status_test.go (279 lines, 15 tests)
  → status_build_test.go, status_session_test.go

- internal/dockerbuild/game_test.go (282 lines, 8 tests)
  → game_preamble_test.go, game_script_test.go

- internal/wsl/detect_test.go (254 lines, 11 tests)
  → detect_distro_test.go, detect_utils_test.go, detect_scripts_test.go

Also updated CLAUDE.md: added DDC architecture section and documented
scripts/validate_ue_versions.sh in the release process description.

All tests pass. Lint clean. gocyclo -over 8 returns nothing.
